rag_web_app/backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9

# RAG stack
faiss-cpu==1.7.4
sentence-transformers==3.0.1
transformers==4.44.2
torch==2.4.0

# PDF
PyPDF2==3.0.1

# CORS utilities
starlette==0.38.4
